Add files via upload
This is used to calculate the value of pi in Julia. 
Given a square of length 2r, the square's area is:
A_{square} = (2r)^2 = 4r^2
whereas the area of a circle with radius r is:
A_{circle} = pi r^2

Combining both formulas, we get:
pi= 4(M/N)

Where N is the random points in the square.
M is the number of the N random points that fell inside the circle.